+++ x/cwregistry/types/msg.go
@@ -1,0 +1,9 @@
+package types
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+var (
+	_ sdk.Msg = &MsgRegisterCode{}
+)